New Env "test" + Add Service Account and permission to create key to it for people (#119)
* we do not version raw csv files

* needed to get the pipeline to work

* BQ tables aren't versioned in kedro way

* filter null

* fixed logging size, avoid parallelism

* parallel to 5

* conc 50

* decrease batch size

* reduce to 500

* do not rerun if embeddings are done

* add more learned errors

* alternative way of reading nodes that is faster with repartition

* Add SA that can read storage and give access to it, fixes #75

* update docs with new SA

* remove versioned:true from BQ DS

* comment to make release-name easier

* comments added

* add kedro cli file to overwrite cli params

* improved docs for matrix pipeline

* move versions to base

* add tags to enable filtering

* introduce new full_local env

* wipe script update to take parameter

* rm spark config change

* move schema validation to base&layer info for viz

* assert not 0

* name category description order

* remove headers from the wrong places in catalog

* give main takeaway for docs onboarding

* turn off verbose enterprise logging

* move prod params to base to simplify

* make fabricator use pyspark to make the schema validation work

* move code to base

* less data fabricated more nodes not used in integration test

* revert version bump

* only 2440 embeddings and then quit

* Update pipelines/matrix/conf/prod/globals.yml

* Update pipelines/matrix/conf/prod/globals.yml

* fix release

* add

* move from full_local to test env

* add env.tmpl file

* add docs

* fix release

* works

* cleanup

* bump after PR ocmments

* not drop automatically

* Update pipelines/matrix/conf/base/embeddings/catalog.yml


* Update pipelines/matrix/conf/test/embeddings/parameters.yml


* Update docs/src/onboarding/index.md


* Apply suggestions from code review


* Update pipelines/matrix/Makefile


---------

+++ pipelines/matrix/scripts/wipe_neo4j.py
@@ -1,5 +1,6 @@
 """Wipes neo4j instance locally."""
 from neo4j import GraphDatabase
+import typer
 from dotenv import load_dotenv
 import os
 
@@ -15,8 +16,8 @@ def avoid_wiping_prod():
 
 def connect_to_neo4j():
     return GraphDatabase.driver(
-        os.environ["NEO4J_HOST"],
-        auth=(os.environ["NEO4J_USER"], os.environ["NEO4J_PASSWORD"]),
+        "bolt://127.0.0.1:7687",
+        auth=("neo4j", "admin"),
     )
 
 
@@ -25,19 +26,42 @@ def get_user_databases(session):
     return [record["name"] for record in result if record["name"] != "system"]
 
 
-def main():
+def main(
+    db_name: str = typer.Argument(
+        None,
+        help="Name of the database to wipe. If not provided, all user databases will be wiped.",
+    ),
+):
     try:
         with connect_to_neo4j() as driver:
             with driver.session(database="system") as system_session:
-                databases = get_user_databases(system_session)
-                print(f"Found {len(databases)} user database(s)")
-                print(databases)
+                if db_name:
+                    confirm = input(
+                        f"Are you sure you want to drop database '{db_name}'? Type 'y' to confirm: "
+                    )
+                    if confirm.lower() != "y":
+                        print("Operation cancelled.")
+                        return
+                    print(f"Dropping database: {db_name}")
+                    system_session.run(f"DROP DATABASE `{db_name}` IF EXISTS")
+                    print(f"Database {db_name} has been wiped.")
+                else:
+                    databases = get_user_databases(system_session)
+                    print(f"Found {len(databases)} user database(s)")
+                    print(databases)
 
-                for db in databases:
-                    print(f"dropping {db}")
-                    system_session.run(f"DROP DATABASE `{db}` IF EXISTS")
+                    confirm = input(
+                        "Are you sure you want to drop all user databases? Type 'y' to confirm: "
+                    )
+                    if confirm.lower() != "y":
+                        print("Operation cancelled.")
+                        return
 
-            print("\nAll user databases have been wiped.")
+                    for db in databases:
+                        print(f"Dropping {db}")
+                        system_session.run(f"DROP DATABASE `{db}` IF EXISTS")
+
+                    print("\nAll user databases have been wiped.")
 
     except Exception as ex:
         print(f"An error occurred: {ex}")
@@ -46,4 +70,4 @@ def main():
 if __name__ == "__main__":
     avoid_wiping_prod()
     load_dotenv()
-    main()
+    typer.run(main)
